requirements.txt
beautifulsoup4==4.13.4
blinker==1.9.0
bs4==0.0.2
click==8.2.1
feedparser==6.0.11
Flask==3.1.1
Flask-WTF==1.2.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
peewee==3.18.2
python-dotenv==1.1.1
sgmllib3k==1.0.0
soupsieve==2.7
typing_extensions==4.14.1
Werkzeug==3.1.3
WTForms==3.2.1
